+++ config/splat.pspeu.dra.yaml
@@ -44,7 +44,6 @@ segments:
     subalign: 8
     subsegments:
       - [0x80, c, dra_psp/80]
-      - [0x1E50, c, dra_psp/1E50]
       - [0x33F0, c, dra_psp/33F0] # O4
       - [0x5DA0, c, dra/624DC] # O4
       - [0x60F8, c, dra/628AC] # O4
